ECO-127: rename name of directory from service to services.
