+++ explorer/sdk/src/index.ts
@@ -7,7 +7,6 @@ export type ByteArray = Uint8Array;
 export type DeployHash = ByteArray;
 export type BlockHash = ByteArray;
 
-export { CasperService, BalanceService, GrpcError } from './service';
+export { CasperService, BalanceService, GrpcError } from './services';
 export { Args, Contracts, Keys, Serialization};
-export {encodeBase64, decodeBase64} from 'tweetnacl-util';
-export {base64to16, encodeBase16, decodeBase16} from './lib/Conversions'
+export {base64to16, encodeBase16, decodeBase16, encodeBase64, decodeBase64} from './lib/Conversions'
